refactor: restructure project into MVC pattern
Change name of package from models to model
Create view package and move InputHandler into it
Create controller package and move Race into it

Refactoring application structure based on knowledge about MVC pattern:
- separates the application into different units to handle different concerns
